requirements.txt
addict==2.4.0
certifi
charset-normalizer==3.1.0
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.10.7
fonttools==4.40.0
fvcore==0.1.5.post20221221
huggingface-hub==0.13.3
idna==3.4
importlib-resources==5.12.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.4.2
kiwisolver==1.4.4
lit==15.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdurl==0.1.2
mmcv==2.0.1
mmcv-full==1.7.1
mmengine==0.8.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
portalocker==3.1.1
pycocotools==2.0.10
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
rich==13.4.2
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.7
six==1.16.0
sympy==1.11.1
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.5.0
timm==0.6.13
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
yacs==0.1.8
yapf==0.32.0
zipp==3.15.0
